requirements.txt
asgiref==3.10.0
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.3
Django==5.2.8
django-extensions==4.1
gunicorn==23.0.0
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
pycparser==2.23
pyOpenSSL==25.3.0
qrcode==8.2
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
whitenoise==6.11.0
